+++ app/enums.py
@@ -3,6 +3,8 @@
 from __future__ import annotations
 
 import enum
+from types import ModuleType
+from typing import Callable, NamedTuple
 
 
 __all__ = ["PermissionType", "PermissionScope"]
@@ -17,6 +19,14 @@ class PermissionType(enum.Enum):
     read = enum.auto()
     write = enum.auto()
     delete = enum.auto()
+
+
+class _PSParams(NamedTuple):
+    allow_elem: bool
+    need_elem: bool
+    parent: str | None = None
+    parent_attr: str | None = None
+    query: Callable[[ModuleType], list] | None = None
 
 
 class PermissionScope(enum.Enum):
@@ -34,20 +44,39 @@ class PermissionScope(enum.Enum):
             ``parent`` is set.
     """
     def __new__(cls, allow_elem: bool, need_elem: bool, _parent: str | None,
-                parent_attr: str | None):
+                parent_attr: str | None, query: Callable[[ModuleType], list] | None = None):
         obj = object.__new__(cls)
         obj._value_ = enum.auto()
         obj.allow_elem = allow_elem
         obj.need_elem = need_elem
         obj.parent = getattr(cls, _parent) if _parent else None
         obj.parent_attr = parent_attr
+        obj.query = query
         return obj
 
     def __repr__(self) -> str:
         return f"<{type(self).__name__}.{self.name}>"
 
-    pceen = (True, False, None, None)
-    photos = (False, False, None, None)
-    collection = (True, False, None, None)
-    album = (True, True, "collection", "collection")
-    role = (True, False, None, None)
+    pceen = _PSParams(
+        allow_elem=True,
+        need_elem=False,
+        query=lambda models: models.PCeen.query,
+    )
+    photos = _PSParams(allow_elem=False, need_elem=False)
+    collection = _PSParams(
+        allow_elem=True,
+        need_elem=False,
+        query=lambda models: models.Collection.query,
+    )
+    album = _PSParams(
+        allow_elem=True,
+        need_elem=True,
+        parent="collection",
+        parent_attr="collection",
+        query=lambda models: models.Album.query,
+    )
+    role = _PSParams(
+        allow_elem=True,
+        need_elem=False,
+        query=lambda models: models.Role.query,
+    )